# OOCL.Client/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "OOCL.Api",
		"version": "1.0"
	},
	"paths": {
		"/api/Audio/audios": {
			"get": {
				"tags": [
					"Audio"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/AudioObjInfo"
									}
								}
							},
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/AudioObjInfo"
									}
								}
							},
							"text/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/AudioObjInfo"
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/{guid}/info": {
			"get": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/{guid}/remove": {
			"delete": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "boolean"
								}
							},
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							},
							"text/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/upload": {
			"post": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "copyGuid",
						"in": "query",
						"schema": {
							"type": "boolean",
							"default": true
						}
					}
				],
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								}
							},
							"encoding": {
								"file": {
									"style": "form"
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/{guid}/download": {
			"get": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							},
							"application/json": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							},
							"text/json": {
								"schema": {
									"type": "string",
									"format": "binary"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/{guid}/waveform64": {
			"get": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/AudioData"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioData"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioData"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/{guid}/play/{volume}": {
			"post": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					},
					{
						"name": "volume",
						"in": "path",
						"required": true,
						"schema": {
							"type": "number",
							"format": "float",
							"default": 0.66
						}
					}
				],
				"responses": {
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/{guid}/stop": {
			"post": {
				"tags": [
					"Audio"
				],
				"parameters": [
					{
						"name": "guid",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"format": "uuid"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/AudioObjInfo"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
		"/api/Audio/audios/stopAll": {
			"post": {
				"tags": [
					"Audio"
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"text/plain": {
								"schema": {
									"type": "integer",
									"format": "int32"
								}
							},
							"application/json": {
								"schema": {
									"type": "integer",
									"format": "int32"
								}
							},
							"text/json": {
								"schema": {
									"type": "integer",
									"format": "int32"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"text/plain": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							},
							"text/json": {
								"schema": {
									"$ref": "#/components/schemas/ProblemDetails"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"AudioData": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid",
						"readOnly": true
					},
					"name": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"filepath": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"meta": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"onHost": {
						"type": "boolean",
						"readOnly": true
					},
					"sizeMb": {
						"type": "number",
						"format": "float",
						"readOnly": true
					},
					"pointerHex": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataType": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataStructure": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"width": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"height": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"base64Format": {
						"type": "string",
						"nullable": true
					},
					"base64Image": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					}
				},
				"additionalProperties": false
			},
			"AudioObjInfo": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid",
						"readOnly": true
					},
					"name": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"filepath": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"meta": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"onHost": {
						"type": "boolean",
						"readOnly": true
					},
					"sizeMb": {
						"type": "number",
						"format": "float",
						"readOnly": true
					},
					"pointerHex": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataType": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"dataStructure": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"samplerate": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"channels": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"bitdepth": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"lenght": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"chunkSize": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"overlapSize": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"form": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"stretchFactor": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"playing": {
						"type": "boolean",
						"readOnly": true
					},
					"playbackState": {
						"type": "string",
						"nullable": true,
						"readOnly": true
					},
					"volume": {
						"type": "integer",
						"format": "int32",
						"readOnly": true
					},
					"duration": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"currentTime": {
						"type": "number",
						"format": "double",
						"readOnly": true
					},
					"lastLoadingTime": {
						"type": "number",
						"format": "double"
					},
					"lastProcessingTime": {
						"type": "number",
						"format": "double"
					},
					"errorMessage": {
						"type": "string",
						"nullable": true
					},
					"bpm": {
						"type": "number",
						"format": "float"
					}
				},
				"additionalProperties": false
			},
...
			"ProblemDetails": {
				"type": "object",
				"properties": {
					"type": {
						"type": "string",
						"nullable": true
					},
					"title": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"type": "integer",
						"format": "int32",
						"nullable": true
					},
					"detail": {
						"type": "string",
						"nullable": true
					},
					"instance": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": {}
			}
		}
	}
}